将迷宫题全部函数分文件但是不清楚分这么细是否有问题，以及当出现某函数需要调用另外一个函数是否要将他们放一起
